moved some interfaces to common + added for logger
