+++ packages/api/src/application/weather-scheduler/weather-scheduler.service.ts
@@ -8,7 +8,7 @@ import { WeatherUpdateInterface } from './interfaces/weather-update.abstract';
 export class WeatherSchedulerService {
   constructor(private readonly weatherUpdateService: WeatherUpdateInterface) {}
 
-  @Cron(CronExpression.EVERY_MINUTE)
+  @Cron(CronExpression.EVERY_HOUR)
   private async sendHourlyUpdates() {
     console.log(NotificationSubjects.WEATHER_UPDATES_HOURLY);
     await this.weatherUpdateService.sendUpdates({
