+++ packages/api/src/application/weather/weather.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
+import { CacheAccessor } from 'src/common/interfaces/cache-service.interface';
+import { ProviderLogger } from 'src/common/interfaces/logger.interface';
 import { CacheModule } from 'src/infrastructure/cache/cache.module';
-import { CacheAccessor } from 'src/infrastructure/cache/interfaces/cache-service.interface';
 import { LoggerModule } from 'src/infrastructure/logger/logger.module';
-import { LoggerService } from 'src/infrastructure/logger/logger.service';
 import { OpenWeatherWeatherProvider } from 'src/infrastructure/weather/providers/openweather.provider';
 import { WeatherApiWeatherProvider } from 'src/infrastructure/weather/providers/weatherapi.provider';
 import { WeatherProviderCacheProxy } from 'src/infrastructure/weather/wrappers/weather-cache.proxy';
@@ -19,8 +19,8 @@ import { WeatherProvider } from './interfaces/weather-provider.abstract';
     OpenWeatherWeatherProvider,
     {
       provide: WeatherProvider,
-      inject: [WeatherApiWeatherProvider, OpenWeatherWeatherProvider, LoggerService, CacheAccessor],
-      useFactory: (weatherApiProvider: WeatherApiWeatherProvider, openWeatherProvider: OpenWeatherWeatherProvider, logger: LoggerService, cacheService: CacheAccessor) => {
+      inject: [WeatherApiWeatherProvider, OpenWeatherWeatherProvider, ProviderLogger, CacheAccessor],
+      useFactory: (weatherApiProvider: WeatherApiWeatherProvider, openWeatherProvider: OpenWeatherWeatherProvider, logger: ProviderLogger, cacheService: CacheAccessor) => {
         const decoratedWeatherAPI = new WeatherProviderLoggingDecorator(weatherApiProvider, logger);
         const decoratedOpenWeather = new WeatherProviderLoggingDecorator(openWeatherProvider, logger);
         const cachProxied = new WeatherProviderCacheProxy(decoratedWeatherAPI.setNext(decoratedOpenWeather), cacheService);
