+++ packages/api/src/common/interfaces/logger.interface.ts
@@ -1,0 +1,3 @@
+export abstract class ProviderLogger {
+  abstract logProvider(message: string, executor: string, data: unknown);
+}
