+++ packages/api/src/infrastructure/city/wrappers/city-provider-logging.decorator.ts
@@ -1,10 +1,10 @@
 import { ChainableCityProvider } from 'src/application/city/interfaces/chainable-city.provider';
-import { LoggerService } from 'src/infrastructure/logger/logger.service';
+import { ProviderLogger } from 'src/common/interfaces/logger.interface';
 
 export class CityProviderLoggingDecorator extends ChainableCityProvider {
   constructor(
     private readonly wrapped: ChainableCityProvider,
-    private readonly logger: LoggerService,
+    private readonly logger: ProviderLogger,
   ) {
     super();
   }
