+++ packages/api/src/infrastructure/logger/logger.module.ts
@@ -1,8 +1,14 @@
 import { Module } from '@nestjs/common';
+import { ProviderLogger } from 'src/common/interfaces/logger.interface';
 import { LoggerService } from './logger.service';
 
 @Module({
-  providers: [LoggerService],
-  exports: [LoggerService],
+  providers: [
+    {
+      provide: ProviderLogger,
+      useClass: LoggerService,
+    },
+  ],
+  exports: [ProviderLogger],
 })
 export class LoggerModule {}
