+++ packages/api/src/infrastructure/logger/logger.service.ts
@@ -1,11 +1,12 @@
 import { Injectable } from '@nestjs/common';
+import { ProviderLogger } from 'src/common/interfaces/logger.interface';
 import { localTimestampFormat } from 'src/infrastructure/logger/configs/timezone';
 import { createLogger, format, Logger, transports } from 'winston';
 
 const { combine, timestamp, json, prettyPrint } = format;
 
 @Injectable()
-export class LoggerService {
+export class LoggerService implements ProviderLogger {
   private readonly logger: Logger;
 
   constructor() {
